Seperate timeconsuming tests to the benchmarkTest task
