+++ modApiServer/build.gradle
@@ -4,11 +4,37 @@ ext.moduleName = 'aion.apiserver'
 build.dependsOn.remove("check")
 
 test.dependsOn copyNativeLibsForModuleTests
+unitTest.dependsOn copyNativeLibsForModuleTests
+benchmarkTest.dependsOn copyNativeLibsForModuleTests
 
 sourceSets {
     test {
         resources {
             srcDirs = ['test_resources']
+        }
+    }
+
+    unitTest {
+        resources {
+            srcDirs = ['test']
+        }
+        java {
+            srcDirs = ['test']
+            exclude '**/AccountManagerTest.java'
+            compileClasspath += main.output + test.output
+            runtimeClasspath += main.output + test.output
+        }
+    }
+
+    benchmarkTest {
+        resources {
+            srcDirs = ['test']
+        }
+        java {
+            srcDirs = ['test']
+            include '**/AccountManagerTest.java'
+            compileClasspath += main.output + test.output
+            runtimeClasspath += main.output + test.output
         }
     }
 }
@@ -47,6 +73,14 @@ dependencies {
     testCompile 'junit:junit:4.12'
     testCompile 'org.hamcrest:hamcrest-all:1.3'
     testCompile "org.mockito:mockito-core:2.23.0"
+}
+
+configurations {
+    unitTestCompile.extendsFrom testCompile
+    unitTestRuntime.extendsFrom testRuntime
+
+    benchmarkTestCompile.extendsFrom testCompile
+    benchmarkTestRuntime.extendsFrom testRuntime
 }
 
 test {
