+++ modP2pImpl/build.gradle
@@ -6,6 +6,17 @@ sourceSets {
             srcDirs = ['test/resources']
         }
     }
+
+    unitTest {
+        resources {
+            srcDirs = ['test/resources']
+        }
+        java {
+            srcDirs = ['test']
+            compileClasspath += main.output + test.output
+            runtimeClasspath += main.output + test.output
+        }
+    }
 }
 
 dependencies {
@@ -18,6 +29,11 @@ dependencies {
     testCompile 'junit:junit:4.12'
     testCompile 'org.hamcrest:hamcrest-core:1.3'
     testCompile "org.mockito:mockito-core:2.23.0"
+}
+
+configurations {
+    unitTestCompile.extendsFrom testCompile
+    unitTestRuntime.extendsFrom testRuntime
 }
 
 // Skip unit tests when doing build task; unit tests are all mixed up with
