+++ modUtil/build.gradle
@@ -43,6 +43,38 @@ sourceSets {
     test {
         java.srcDirs = ['src/test/java']
     }
+
+    unitTest {
+        resources {
+            srcDirs = ['src/test/java']
+        }
+        java {
+            srcDirs = ['src/test/java']
+            exclude '**/Benchmark.java'
+            compileClasspath += main.output + test.output
+            runtimeClasspath += main.output + test.output
+        }
+    }
+
+    benchmarkTest {
+        resources {
+            srcDirs = ['src/test/java']
+        }
+        java {
+            srcDirs = ['src/test/java']
+            include '**/Benchmark.java'
+            compileClasspath += main.output + test.output
+            runtimeClasspath += main.output + test.output
+        }
+    }
+}
+
+configurations {
+    unitTestCompile.extendsFrom testCompile
+    unitTestRuntime.extendsFrom testRuntime
+
+    benchmarkTestCompile.extendsFrom testCompile
+    benchmarkTestRuntime.extendsFrom testRuntime
 }
 
 signing {
